Support for batching in click
Elements can explicitely support batching by inheriting BatchElement and
implementing push_batch and/or pull_batch or simple_packet_batch.

Batching is retro-compatible. Ports from and to BatchElement are upgraded
and if they are between a compatible and an incompatible batching element,
they will unbatch/rebatch packets.

For a push_batch to a normal element, batch is recreated in the port of
all downstream batching compatible element. When all packets of the batch
has been pushed, end_batch() is called.

Pull path does not support batching, because the batch taken from upstream
could be splitted on the way back to the downstream elements and produce
unexpected behaviour.