Support pull paths and add need_batch() flag
Support pull_batch for BatchElement. On a pull_batch to a non-batching element, the port makes a batch by calling pull() on the upstream element until it returns null or it reaches MAX_PULL_BATCH (currently 1024).

Also add need_batch() constant bool function which return true if the element needs to always be called by push_batch() and pull_batch() and never by pull() or push(). Therefore, need_batch()should return true if the element does not implement push() or pull().
push() is called on BatchElement when reconstructing the batch would be done only for this element then re-deconstructed again. We tried to avoid recontructing a batch just for one element, but it may be necessary (for example in some some output element) an enforced via need_batch().