Refactorize (From|To)(Netmap|DPDK)Device and add QUEUE/N_QUEUES parameters
Remove duplicate code and highlight common parameters, variables and functions. QueueDevice is also extended in RXQueueDevice and TXQueueDevice

Add a QUEUE parameter to both elements to specify a queue to use instead of using auto-assignation, and a N_QUEUES parameter to fix the numbers of queues to use following the queue index QUEUE instead of using the default behaviour. The default behaviour is to always start at QUEUE 0, and use all of available RX queues with FromNetmap or the best number (=threads assigned to it) with FromDPDKDevice, or as many as there is threads ending up in the given To(DPDK|Netmap) element.

Also rework ToDPDKDevice::push_batch to use more or less the same mechanism than push_packet and also avoid duplicated code. Moreover the last code was introducing reordering.