TSCClock: introduce user-provided clock
If click is compiled with --enable-user-timestamp, an element will be
able to provide a clock function to Timestamp to get current wall clock
and steady clock time instead of using the default system.

TSCClock uses that system to provide the time using the TSC.

At beginning, the system clock is untouched and Click will run normally.
A timer is scheduled at 10HZ and goes through the following different phases.

The first phase allows to compute a correct frequency, and stops when the TSC
clock is close enough to the real clock after much iterations. The clock
then moves to the synchronization phase.

This phase has to goals :
 - Ensure that the time is correct on all threads. If any threads get
10 bad times (error bigger than 100subsec), the phase fails. The clock
may go to the next phase only when 10 good reads happened.
 - Compute any per-core TSC offset, for older CPUs where the TSC can
start with different offset on different cores.

If threads are afinitized to cores (forced with DPDK, -a argument with
userlevel click, --cpu with kernel click, or your own action), the offset
will be kept.
If it is not the case, the thread may change of core and the thread local TSC
offset would not mean anything anymore. If an offset is effectively
needed, and ALLOW_OFFSET is not true (default with DPDK only), this
phase will fail. If you manage your own thread scheduling and are sure
that the thread will always stay on the same core (isolcpus,...) you may
pass ALLOW_OFFSET true.

If any of the previous tests fails, the clock will not be set. If they
pass, it will call the Timestamp::set_clock function when leaving
synchronization phase and from that point
the clock will really be used. This takes around ~15 seconds. Both wall
and steady clocks are based on the system times, so the transition is
seamless.

After that, as the sytem and NTP are applying correction to the wall clock,
the timer continues to tick to change a little bit the frequency so the
clock can catch up with system clock.

If the drift is too big (NTP update, probably), the clock will jump.
This already happens with system clock time because of NTP, so nothing new
here . In other situation, the frequency will be
slightly modified so the clock catch up with system time, without
jumping.

The steady clock frequency will be fixed after the stabilisation phase
and will never change. The frequency could be off by a few hertz, but it
is not a problem as the steady clock should not be used for long
periods. And when it is (timers) a few nanoseconds shift every seconds
is not a big problem. The steady clock stills need to be accumulated as
if the delta ticks is bigger than 2^32 (less than 2sec on a 3ghz
computer), the computation will overflow. This case will however be detected by
the update tick. We will move the timer to the next thread, as there may
simply be a long task on the current thread. If we do a full turn, the
clock is deactivated...

Your system must support nonstop_tsc and invariant_tsc cpu
flags to use the TSC clock. In any case, do some testing before using this
element. Any recent Intel CPU should be fine though. Normally, the
different checks will ensure that those are setted, as the
synchronization step will never passthrough if those flag are not there.

To use the feature, compile with --enable-user-timestamp and place the
TSCClock element without any input or output in your configuration.
The TSCClock will add a timer which will tick at 10HZ. Assign this
element to a thread wich will not suffer from this, and which will not
have task longer than one second. If heavy tasks preven the timer to run
at 10HZ, it will try all other threads before deactivating the clock.