Re-design of batching
I now use only the ports to do re-batching and de-batching.
This allows to get rid of the push_packet change, and also
have de-batching and re-batching per-port instead of only
the port 0.

This makes the implementation much cleaner, I changed the base
port Element so there is no need for upgrade_ports() anymore.

Technically, a batch-compatible element does not even need
anymore to extend BatchElement, it could simply set
in_batch_mode to BATCH_MODE_YES if it always produces batches,
or BATCH_MODE_IFPOSSIBLE if it will do like the input do (eg
push if push is called, pull if pull is called).

But BatchElement stays there for convenience, it is probably
good practice to differenciate element with proper support.

Secondly, this patch allows for automatic batching ! Hurra.
When a push_batch is called on a normal element,
all its output port will buffer packets. push_batch then call
push for each packets, and then all output ports output
the buffer packets as batch in one step.

Use --enable-auto-batch to enable this. It is still a little
experimental so it is not default. Another way to force the
behaviour without global change is to set in_batch_mode to
BATCH_MODE_IFPOSSIBLE in the constructor of vanilla elements
one by one.

However the batching bridge feature may be a little slower,
that i be the ability to jump multiple vanilla elements
and rebuild far away from the de-batching port, not re-batching
after each vanilla element. This patch will not do
a traversal to get a list of the downwards element at configuration
time anymore, it will call the function on each outut port
at run time, which will themselves call the function for the next
ports and so one. However I expect AUTO_BATCH to become the standard
and let this functionnality completely disappear.