Give more clear names to get_threads and get_runnable_threads
get_threads now becomes get_passing_threads, as it returns
the threads pushing and/or pulling to the given element. It
should not be re-implemented.

get_runnable_threads becomes get_spawining_threads as it sets
the bit of the threads that this element will spawn. It must
be reimplemented in non-obvious cases, see Element::get_spawining_threads.