fix performance regression with ToDPDKDevice
Allow to use burst with Batching in ToDPDKDevice, but by default set
it to 1 (push any batch incoming directly to the DPDK ring), and when the
batch is set 1, disable the timeout (as the timeout will always flush nothing).

Problem was that the useless timeout in this case created more contention on the
PCI bus, leading to worst performance when using batching.