Always set timer if packets are awaiting
This commit set the timer if some packets are awaiting in the internal queue,
no mater the timeout.

We also disable the timeout if burst is 1 and timeout is the default (0).
The effect is that the timer will still be setted to run after all task have completed but only if there is pending packets in the queue after sending, instead of putting it anyway even if the internal queue is empty (which is useless).

And fix bug where the iqueue is modified outside the lock