Fix bug where the iqueue is modified outside the lock
Also only disable timeout if burst is 1 and timeout is the default (0).
The effect is that the timer will still be setted to run after all task have completed but only if there is pending packets in the queue after sending, instead of putting it anyway even if the internal queue is empty (which is useless).