Allow replay of packets through RecordTimestamp
This allows packets to be replayed through the recordtimestamp ecosystem
by setting the OFFSET argument in RecordTimestamp. That will force record timestamp to read the number inside the packet to find the timestamp position.

Note that this will only work if there are more packets to replay
than the maximal amount of packets in the total trip path. Indeed, if
you replay only a few packets, the timestamp will be overwritten before
it is actually read by TimestampDiff.