Dynamic compilation of flow director library
This patch gets rid of the need to maintain testpmd.c in Click
in order to compile librte_parse. Instead, we dynamically copy
and process testpmd.cduring the compilation process. This patch
also ensures correct compilation of FastClick with different
DPDK versions that might or might not support librte_parse.

Signed-off-by: Georgios Katsikas <katsikas.gp@gmail.com>