New socket element for monitoring purposes
This element acts as a UDP or TCP client that reports
a list of handlers' values and names to a remote UDP or TCP
server. Currently, this element is useful for element handlers
that contain arithmetic values (e.g., latency, throughput, etc).
The structure of each message is <timestamp, handler value, handler name>.

Signed-off-by: Georgios Katsikas <katsikas.gp@gmail.com>