Updates and bug fixes in TimestampDiff
This patch contains re-worked statistics' calculation
methods, 2 bug fixes, and a new handler providing the last
latency value in the buffer.

One bug was related to the standard deviation calculation.
If variance was zero, then sqrt(0) was problematic.
The second bug was related to the min handler. If no
traffic is present, then this counter was MAX_UINT.
Now this is replaced by zero.

Signed-off-by: Georgios Katsikas <katsikas.gp@gmail.com>