Extensions in rule generation elements
Base element GenerateIPFilter can now be configured with pattern type.
This type can be IPFILTER or IPCLASSIFIER for GenerateIPFilter, IPLOOKUP for
GenerateIPLookup, or FLOW_DIRECTOR for GenerateIPFlowDirector.

A bug was also fixed in GenerateIPLookup. That bug was preventing to output
the output port of each rule, due to missing type casting.

Signed-off-by: Georgios Katsikas <katsikas.gp@gmail.com>