[Metron] Extended and correct rule management
This patch complies with the reworked Flow Director API
and introduces additional functionality in Metron's NIC
rule management.
First, rule insertion follows a two phase commit process
which ensures that a new rule is inserted after deleting
its previous occurence (which might be mapped to another
CPU core). This might occur during a load balancing, when a
rule was initially mapped to e.g., core 0 but now needs to
be mapped to another core.

Rule deletion is tricky because rule IDs as advertized by the
controller are different from rule IDs as understood by NICs.
To solve this problem, an internal map is implemented to keep
the mappings between ONOS rule IDs and DPDK NIC rule IDs.

Signed-off-by: Georgios Katsikas <katsikas.gp@gmail.com>