Fix Kernel compilation up to at least 4.4
And move things around to remove things uncompatible with kernel mode.

Among them batching. But it is not a technical problem, it justs needs a
little refactoring to work in kernel.

Anyway, as mainline Click, the kernel mode is not working very well with
recent versions, that should be fixed before batching.

There may be some unfixed platform dependencies such as ARM, don't
hesitate to fix them too.