Fix the threading model
Use lambda to be able (in future) to delay the execution of a script
until the change is really made.
Also change the message to separate UP and DOWN events, allowing to
correctly handle one or the other.

For instance, ToDPDKDevice wants to deallocate tasks AFTER some thread
are disabled, while it wants to allocate them BEFORE some thread are
going to get through us.