Rewrite CPU organization
Have all per-cpu things of the service chain in CpuInfo.
Explicitely mention when a CPU id is a physical ids. Eg a SC with 3 cpus
will always have cpu id from 0 to 2 but that may mayb to cores 7 9 and
10, which are the physical id.
Separate clearly what is "allocated cpus" (max number of CPU that are
allocated on physical cpus) and "active cpus" that are the CPU up by the
controller, and that are actually polling packets.