Cleaner integration between Metron and Flow Director
This patch moves some of the NIC functionality previously in
Metron to Flow Director. This makes Flow Director a single
point for managing NIC resources. To facilitate these changes,
Flow Director is extended with a Flow Cache, which keeps important
data structures for rule management and monitoring.

Also, Metron and FromDPDKDevice are extended with additional
handlers to obtain further information from Flow Director.

Signed-off-by: Georgios Katsikas <katsikas.gp@gmail.com>