lib/packet: add option to make packet allocations fail 10% of the time
This change adds--enable-flaky-packets, which causes Packet methods that
might allocate to fail 10% of the time.  This includes Packet::make,
Packet::clone, and Packet::expensive_uniqueify.  Because Packet::put &
Packet::push use uniqueify to allocate, it covers those expensive operations
as well but only in the case that there is not enough headroom/tailroom in
the packet to complete without allocating.  This avoids false positives,
e.g.  when the caller knows that the opearation must succeed because
Packet::make was invoked with sufficient headroom/tailroom.

With valgrind-helper, this option is useful for finding callers that do not
deal with allocation failure correctly.

*** NB: test-click will fail all over the place, since this change eats SKBs

Change-Id: Id36175436ae7006b54c3590878988ec0a94f5166
Signed-off-by: Derrick Pallas <pallas@meraki.com>