click/fromfile: plug mmap leak in error path
If mmap is successful but Packet::make fails to create a new SKB, we need to
execute the destructor we passed to Packet::make.  It actually might make
sense to do this in Packet::make instead instead of the caller, not sure
which is better.

Change-Id: If97d221daf3443003f40097cc36f7410bb544669
Reviewed-on: https://gerrit.ikarem.io/74137
Static-Analysis: Jenkins Build System <jenkins@meraki.com>
Tested-by: Jenkins Build System <jenkins@meraki.com>
Reviewed-by: Peter Hurley <peter@meraki.com>
Signed-off-by: Derrick Pallas <pallas@meraki.com>