click/hash*: add hash-iterator-epochs configure option
This option causes HashContainer to track an epoch value, which is bumped
when iterator-invalidating methods are invoked.  When iterators are created
or updated, they take the current epoch.  When dereferenced or otherwise
used, iterators now assert that their epoch matches the epoch of the
HashContainer.

This allows one to find instances of bad iterator use, e.g.

	for (Table::iterator i = table.begin() ; i ; ++i)
		if (predicate)
			table.erase(i);

This has sometimes been "fixed" by changing the code to say

	for (Table::iterator i = table.begin() ; i ; )
		if (predicate) {
			Table::iterator tmp = i;
			++i;
			table.erase(tmp);
		} else
			++i;

but because these iterators are fat, erase still invalidates i and later use
will fail sporadically.  The correct fix is

	for (Table::iterator i = table.begin() ; i ; )
		if (predicate)
			i = table.erase(tmp);
		else
			++i;

since erase takes care to return a now-valid iterator.  This error does not
just apply to erase, but anything that might modify the linkage within the
hashtable, including find_insert, find_prefer, rehash, &c.

Thanks to Gjergji Zyba for identifiying the root issue in HashContainer.

Change-Id: I9a8e8f9c72d4483c2840f8901e78ea8fcbb3a141
Signed-off-by: Derrick Pallas <pallas@meraki.com>