Metron: Mirror mode
Will automatically reflect mirror ruled on the second port of each pair
of port. Eg if chains have fd0 and fd1, then a rule installed of fd0
such as "src ip A, dst ip B" will also lead to the installation of
"dst ip A, src ip B" on fd1.