Support DPDK 19.05
Also switch to meson when available. In the future DPDK plan to switch
to meson and we will use pkg-config to get dependencies.

This allows to remove the ugly code of dpdk.mk. It will be removed when
DPDK will fully move to meson.

Parallel development with Georgios Katsikas <katsikas.gp@gmail.com>

Signed-off-by: Tom Barbette <barbette@kth.se>