LRO, Rx jumbo frames, and VLAN extend offloading
To allow Rx jumbo frames offloading it is necessary
to set max_rx_pkt_len, which was not set by FastClick.
Also, fixed the initial VLAN filter and strip configuration
status (which was mistakenly set to true).
Messages are added to notify users about the status of
the offloading features when turned on.

Signed-off-by: Georgios Katsikas <katsikas.gp@gmail.com>