Solve a huge source of cache misses
The "lock" mechanism was supposed to be enabled only if queues had to be
shared. To do so, the value of the lock would be "2" (NO_LOCK) and would
never change to mean "disabled". However as that number is atomic, it
was re-read every time. Moreover, those per-thread values were not
memory-aligned, leading to C2C transfers at the same time. A huge mess.

To do this we also allow an aligned version of Vector, the default tries
to be conservative, but one can pass a specific alignement to Vector,
such as CLICK_CACHE_LINE_SIZE in this case