Ability to load millions of flow rules
The current stack allocation of flow rule IDs
limits the ability to store huge batches of rules.
As some NICs (e.g., Mellanox) allow to store millions
of rules, we updated the process of adding rules from
file, to scale the number of istalled rules to several
millions.

Documentation updates and refactoring are also performed.

Signed-off-by: Georgios Katsikas <katsikas.gp@gmail.com>