String searching element and library
The Search() element allows to search for a pattern in a packet. It will
then Strip to the beginning (STRIP_AFTER false) or the end of the pattern (default, intended
to skip the payload to delimiters such as "\n\r\n\r" in HTTP headers).
Search also sets the new PAINT2 annotation (like Paint, but 2 bytes) to the amount of skipped
bytes, so one can go backwards after a search using UnstripAnno.

The underlying lib is using a new search function of the String class,
which will use AVX2 lookup if possible (detected via a new configure
line).