atomics: allow to use C++11 implementation
C++11 atomics can be used on platforms for which the atomic
implementation of click is not given (e.g. ARM).
To enable the new implementation, use the configuration flag
`--enable-atomic-builtins`.

compare_and_swap bool version is preferred if compiled with builtins,
otherwise the int version will be used. This is controlled by the
CLICK_ATOMIC_COMPARE_SWAP definition.