Meson: fix in version check and librte_parse compilation
When compiling using meson, the RTE_SDK path should be set, otherwise it
is impossible to find the testpmd source folder and compile the parsing
library.

The compilation of librte_parse library has been separated from the
dpdk.mk file so it can be included without defining all the dpdk
variables.

Fix in RTE_VER variables: month and minor version is set also when DPDK
is found with pkg-config.