Script: Add gotoa instruction / goto "alive"
When a driver terminates, the goto instruction would repeat itself until
it reaches the limit of 1000 max jumps in a row. When printing stuff as
in a cron job, it would produce a thousand print when killing the
router, and stop the script there. Goto "while alive" will stop
gotoing when the router is trying to shutdown, allowing to have a
gracefull shutdown.