Help LTO inliner with hints
Somehow, in previous tests xchg_ functions were automatically inlined.
In some recent tests they were not. So instead of relying on luck, we'll
add the ALWAYS_INLINE attribute that is kept up to LLD, allowing LLD
to inline those functions that do one or two instructions and that
should definitely be inlined.