Add ELEMENT_FEATURES
Indicates when an element should be spawned to populate another.

The use case is for instance FromIPSummaryDump, which uses the
IPSummaryDump features. A lot of elements like IPSummaryDump_TCP
will populate the fields known to IPSummaryDump.

Now when one wants to compile a mindriver, which uses the IPSummaryDump
features, one would expect to have the same behavior, that is also
including all those elements that populates IPSummaryDump.