Merge the second part of MiddleClick (#343)
See https://ieeexplore.ieee.org/abstract/document/9503101?casa_token=E-zGWjVaVIYAAAAA:igHhzDxGujzPkztw4dTBuRUTjoYmp6ZRRl6h6dMMhN0M_5OWZL-0XuoeiQ_d4DD05PC7FZpvYA for the full paper about "MiddleClick"

The flow subsystem has already been merged, that was :

State factorization : the ability for elements to extend
FlowSpaceElement. Such elements will receive packets via a push_flow
function, that on top of the packets of the same session will have a pointer
for an Object* space implemented per-session.
A serie of FlowIPManager elements that will collect the space needed
for all those FlowSpaceElements and classify sessions of packets,
(pre)-allocating the space for all those elements in a single
classification step. The table will be sized to have enough space for
all kind of Object size in the path, without the need for dynamic
allocation. Alternatively elements can can ask for a void* and handle
memory themselves.
This commits merges the rest of the MiddleClick paper's novelties:

The ability for the CTXManager to collect "context", that is classification
rules such as 14/0800, that the CTXManager will manage, like the sessions for the
FlowIPManager but it works for any session definition, not just the 4 tuples, and
any static classification. The classification will be minimized. Eg
classifying in a path for " IP -> Destination -> IP -> Port -> 4-tuple
session" will lead to "IP -> Destination -> Port -> 2-tuple session".
The new "~>" context link that will spawn classification rules
according to the context of elements. For instance adding ~> IPIn
is the same as adding -> FlowDispatcher(12/0800, - drop) -> IPIn. It
can support multiple ports too.
A serie of context entry/exit elements such as TCPIn/TCPOut,
HTTPIn/HTTPout as presented in the paper to allow modification of bytestream.
Elements working on context implement CTXSpaceElement, without
terminating the TCP connection.
A multitude of unit tests for theses.