WIP : Merge interesting results of the Massimo Girondi, Marcho Chiesa and myself conntrack paper
Basically the default FlowIPManager/IMP is now be the Cuckoo++ version, with the timerwheel for recycling.

This commit does not include robinscotch and hopscotch as the paper found them slower in all cases.
Similarly, we removed the support for lazy deletion in the modified DPDK, as it would need to be merged by DPDK people, and it was also found to be unnecessary for performance.