Add basic BIER elements
This commit adds basic elements for BIER packets processing.

First, it adds the Bier Forwarding Table (BIFT) which is implemented as the IPv6 forwarding table.
That is, Fastclick is extended with a 'biertable' that defines the basic operations on the BIFT.
The 'bierroutetable' element defines handlers allowing external interactions and the
'lookupbiertable' is the element representing the BIFT.

Second, two simple filtering elements are added, (i) 'checkbierheader' that verifies the header of
a BIER packet and (ii) 'checkbierin6header' that filters IPv6 packet to detect BIERin6 encapsulated
BIER packets.